backend/requirements.txt
# Core
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0
sentry-sdk==2.8.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Celery
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1

# HTTP (webhook delivery)
requests==2.32.3

# Testing
pytest==8.2.2
pytest-django==4.8.0
pytest-xdist==3.5.0
factory-boy==3.3.0

#Docs
drf-spectacular==0.29.0